+++ hash_table7.hpp
@@ -1378,7 +1378,7 @@ public:
         //if (sizeof(KeyT) < sizeof(size_type) && buckets > (1ul << (sizeof(uint16_t) * 8)))
         //    buckets = 2ul << (sizeof(KeyT) * 8);
 
-        assert(buckets < max_size() && buckets > _num_filled);
+        assert(buckets < (uint64_t)max_size() && buckets > (uint64_t)_num_filled);
         //TODO: throwOverflowError
 
         auto num_buckets = (size_type)buckets;
